feat: rename the mod
