+++ src/lib.rs
@@ -1,5 +1,4 @@
 pub mod challenge;
-pub mod compute;
 pub mod compute_fej;
 pub mod compute_r;
 pub mod inversion;
@@ -8,5 +7,6 @@ pub mod pairing;
 pub mod proof;
 pub(crate) mod serde;
 pub mod test;
+pub mod utils;
 pub mod verifier;
 pub mod vk;
